+++ EdpOnenet/SYSTEM/dht11/dht11.h
@@ -12,7 +12,7 @@
 void DHT11_IO_OUT(void);//温湿度模块输出函数
 void DHT11_IO_IN(void); //温湿度模块输入函数
 void DHT11_Init(void);  //初始化DHT11
-u8   DHT11_Read_Data(u8 *temp,u8 *humi);//读取温湿度
+u8   DHT11_Read_Data(int *temp,int *humi);//读取温湿度
 u8   DHT11_Read_Byte(void);             //读出一个字节
 u8   DHT11_Read_Bit(void);              //读出一个位
 u8   DHT11_Check(void);                 //检测是否存在DHT11
